+++ client/components/pages/submission-system/modals/editUser/index.js
@@ -4,7 +4,7 @@ import Modal from 'uikit/Modal';
 import { UserSection } from '../styledComponents';
 import { UserModel } from '../common';
 import useFormHook from 'global/hooks/useFormHook';
-import { editUserSchema } from '../validations';
+import editUserSchema from './validation';
 
 const EditUserModal = ({
   user,
