+++ client/components/pages/submission-system/modals/editUser/validation.js
@@ -1,0 +1,8 @@
+import * as yup from 'yup';
+import { role } from 'global/utils/form/validations';
+
+const editUserSchema = yup.object().shape({
+  role,
+});
+
+export default editUserSchema;
